minor fixes :D
