+++ src/test/java/pages/LoginPage.java
@@ -1,9 +1,9 @@
 package pages;
 
+import com.codeborne.selenide.SelenideElement;
+
 import pages.main.MainPage;
 import utils.User;
-
-import com.codeborne.selenide.SelenideElement;
 
 import static com.codeborne.selenide.Selectors.byXpath;
 import static com.codeborne.selenide.Selenide.$;
