+++ src/test/java/tests/group/SubUnsubGroupUserViewTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import pages.group.GroupPage;
 import pages.LoginPage;
 import pages.group.UserGroupsPage;
 import tests.BaseTest;
