+++ src/test/java/tests/music/MusicSearchTest.java
@@ -1,14 +1,17 @@
 package tests.music;
 
-import org.junit.jupiter.api.Timeout;
-import pages.LoginPage;
-
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Timeout;
+
+import pages.LoginPage;
 import tests.BaseTest;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static tests.music.SongConstants.*;
+import static tests.music.SongConstants.FIRST_GROUP_NAME;
+import static tests.music.SongConstants.FIRST_SONG_NAME;
+import static tests.music.SongConstants.SECOND_GROUP_NAME;
+import static tests.music.SongConstants.SECOND_SONG_NAME;
 
 public class MusicSearchTest extends BaseTest {
     // Тест: логинимся -> идем в окно с музыкой -> ищем в поиске песню -> проверяем, что она есть
