+++ src/test/java/tests/photo/CreatingAlbumTest.java
@@ -1,6 +1,5 @@
 package tests.photo;
 
-import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
